Move statistics tests to separate file
