+++ sdks/java/extensions/sql/src/test/java/org/apache/beam/sdk/extensions/sql/meta/provider/kafka/BeamKafkaTableTest.java
@@ -17,8 +17,6 @@
  */
 package org.apache.beam.sdk.extensions.sql.meta.provider.kafka;
 
-import java.util.List;
-import org.apache.beam.sdk.extensions.sql.impl.BeamTableStatistics;
 import org.apache.beam.sdk.testing.PAssert;
 import org.apache.beam.sdk.testing.TestPipeline;
 import org.apache.beam.sdk.transforms.Create;
@@ -27,19 +25,12 @@ import org.apache.beam.sdk.transforms.SimpleFunction;
 import org.apache.beam.sdk.values.KV;
 import org.apache.beam.sdk.values.PCollection;
 import org.apache.beam.sdk.values.Row;
-import org.apache.beam.vendor.guava.v26_0_jre.com.google.common.collect.ImmutableList;
-import org.junit.Assert;
 import org.junit.Rule;
 import org.junit.Test;
 
 /** Test utility for BeamKafkaTable implementations. */
 public abstract class BeamKafkaTableTest {
   @Rule public TestPipeline pipeline = TestPipeline.create();
-
-  protected static final List<String> TOPICS = ImmutableList.of("topic1", "topic2");
-
-  /** Returns proper implementation of KafkaTestTable for the tested format. */
-  protected abstract KafkaTestTable getTestTable(int numberOfPartitions);
 
   /** Returns proper implementation of BeamKafkaTable for the tested format. */
   protected abstract BeamKafkaTable getBeamKafkaTable();
@@ -49,87 +40,6 @@ public abstract class BeamKafkaTableTest {
 
   /** Provides a deterministic row from the given integer. */
   protected abstract Row generateRow(int i);
-
-  @Test
-  public void testOrderedArrivalSinglePartitionRate() {
-    KafkaTestTable table = getTestTable(1);
-    for (int i = 0; i < 100; i++) {
-      table.addRecord(createKafkaTestRecord("k" + i, i, 500L * i));
-    }
-
-    BeamTableStatistics stats = table.getTableStatistics(null);
-    Assert.assertEquals(2d, stats.getRate(), 0.001);
-  }
-
-  @Test
-  public void testOrderedArrivalMultiplePartitionsRate() {
-    KafkaTestTable table = getTestTable(3);
-    for (int i = 0; i < 100; i++) {
-      table.addRecord(createKafkaTestRecord("k" + i, i, 500L * i));
-    }
-
-    BeamTableStatistics stats = table.getTableStatistics(null);
-    Assert.assertEquals(2d, stats.getRate(), 0.001);
-  }
-
-  @Test
-  public void testOnePartitionAheadRate() {
-    KafkaTestTable table = getTestTable(3);
-    for (int i = 0; i < 100; i++) {
-      table.addRecord(createKafkaTestRecord("1", i, 1000L * i));
-      table.addRecord(createKafkaTestRecord("2", i, 500L * i));
-    }
-
-    table.setNumberOfRecordsForRate(20);
-    BeamTableStatistics stats = table.getTableStatistics(null);
-    Assert.assertEquals(1d, stats.getRate(), 0.001);
-  }
-
-  @Test
-  public void testLateRecords() {
-    KafkaTestTable table = getTestTable(3);
-
-    table.addRecord(createKafkaTestRecord("1", 132, 1000L));
-    for (int i = 0; i < 98; i++) {
-      table.addRecord(createKafkaTestRecord("1", i, 500L));
-    }
-    table.addRecord(createKafkaTestRecord("1", 133, 2000L));
-
-    table.setNumberOfRecordsForRate(200);
-    BeamTableStatistics stats = table.getTableStatistics(null);
-    Assert.assertEquals(1d, stats.getRate(), 0.001);
-  }
-
-  @Test
-  public void testAllLate() {
-    KafkaTestTable table = getTestTable(3);
-
-    table.addRecord(createKafkaTestRecord("1", 132, 1000L));
-    for (int i = 0; i < 98; i++) {
-      table.addRecord(createKafkaTestRecord("1", i, 500L));
-    }
-
-    table.setNumberOfRecordsForRate(200);
-    BeamTableStatistics stats = table.getTableStatistics(null);
-    Assert.assertTrue(stats.isUnknown());
-  }
-
-  @Test
-  public void testEmptyPartitionsRate() {
-    KafkaTestTable table = getTestTable(3);
-    BeamTableStatistics stats = table.getTableStatistics(null);
-    Assert.assertTrue(stats.isUnknown());
-  }
-
-  @Test
-  public void allTheRecordsSameTimeRate() {
-    KafkaTestTable table = getTestTable(3);
-    for (int i = 0; i < 100; i++) {
-      table.addRecord(createKafkaTestRecord("key" + i, i, 1000L));
-    }
-    BeamTableStatistics stats = table.getTableStatistics(null);
-    Assert.assertTrue(stats.isUnknown());
-  }
 
   @Test
   public void testRecorderDecoder() {
@@ -162,9 +72,5 @@ public abstract class BeamKafkaTableTest {
     public KV<byte[], byte[]> apply(byte[] bytes) {
       return KV.of(new byte[] {}, bytes);
     }
-  }
-
-  private KafkaTestRecord createKafkaTestRecord(String key, int i, long timestamp) {
-    return KafkaTestRecord.create(key, generateEncodedPayload(i), "topic1", timestamp);
   }
 }
